Shift tasks into separate package
