+++ src/main/java/duke/Parser.java
@@ -1,5 +1,10 @@
 package duke;
 
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Task;
+import duke.tasks.Todo;
+
 import java.time.LocalDate;
 
 /**
