+++ src/main/java/duke/Storage.java
@@ -1,5 +1,9 @@
 package duke;
 
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Task;
+
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
