+++ src/main/java/duke/TaskType.java
@@ -1,5 +1,10 @@
 package duke;
 
+import duke.tasks.Deadline;
+import duke.tasks.Event;
+import duke.tasks.Task;
+import duke.tasks.Todo;
+
 /**
  * Enum representing the type of a task.
  */
