+++ src/main/java/duke/tasks/Event.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.tasks;
+
+import duke.tasks.Task;
 
 import java.time.LocalDate;
