+++ src/main/java/duke/tasks/Task.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.tasks;
 
 /**
  * Abstract class representing a task.
