+++ src/main/java/duke/tasks/Todo.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.tasks;
+
+import duke.tasks.Task;
 
 /**
  * Class for Todo tasks.
