+++ src/test/java/ParserTest.java
@@ -4,12 +4,12 @@ import java.time.LocalDate;
 
 import org.junit.jupiter.api.Test;
 
-import duke.Deadline;
-import duke.Event;
+import duke.tasks.Deadline;
+import duke.tasks.Event;
 import duke.Parser;
-import duke.Task;
+import duke.tasks.Task;
 import duke.TaskType;
-import duke.Todo;
+import duke.tasks.Todo;
 
 public class ParserTest {
     @Test
